refactor: move test module to smtp
